refactor(reader): 提取 TTS 页面组件到独立模块
