+++ packages/app-expo/src/components/reader/TTSPage.tsx
@@ -10,11 +10,9 @@ import {
   SkipForwardIcon,
   SquareIcon,
 } from "@/components/ui/Icon";
-import { type ThemeColors, fontWeight, radius, useColors, withOpacity } from "@/styles/theme";
+import { useColors, radius } from "@/styles/theme";
 import {
   buildNarrationPreview,
-  DASHSCOPE_VOICES,
-  EDGE_TTS_VOICES,
   type TTSConfig,
   type TTSPlayState,
   getTTSVoiceLabel,
@@ -23,225 +21,22 @@ import { useCallback, useEffect, useMemo, useRef, useState } from "react";
 import { useTranslation } from "react-i18next";
 import {
   ActivityIndicator,
-  Dimensions,
-  Image,
   Modal,
   Platform,
   Pressable,
   ScrollView,
-  StyleSheet,
   Text,
   TouchableOpacity,
   View,
 } from "react-native";
 import { useSafeAreaInsets } from "react-native-safe-area-context";
+import { BookCoverImage } from "./tts/BookCoverImage";
+import { makeStyles, SH } from "./tts/tts-page-styles";
+import { VoicePickerModal } from "./tts/VoicePickerModal";
 
-// ── Dimensions ────────────────────────────────────────────────────────────────
-const { height: SH, width: SW } = Dimensions.get("window");
-// Full-size cover — 28:41 book aspect ratio
-const COVER_H = Math.round(SH * 0.375);
-const COVER_W = Math.round(COVER_H * (28 / 41));
-// Thumbnail for lyrics header
+// ── Local constants ───────────────────────────────────────────────────────────
 const THUMB_W = 48;
 const THUMB_H = Math.round(THUMB_W * (41 / 28)); // ≈ 70
-
-// ── BookCoverImage — pure render helper ───────────────────────────────────────
-
-interface BookCoverImageProps {
-  coverUri?: string;
-  bookTitle: string;
-  chapterTitle: string;
-  width: number;
-  height: number;
-  borderRadius: number;
-  pct: number;
-  colors: ThemeColors;
-  t: (key: string) => string;
-}
-
-function BookCoverImage({
-  coverUri,
-  bookTitle,
-  chapterTitle,
-  width,
-  height,
-  borderRadius,
-  pct,
-  colors,
-  t,
-}: BookCoverImageProps) {
-  const showStrip = pct > 0 && pct < 100;
-  const fontSize = Math.max(8, Math.round(width * 0.055));
-  const subFontSize = Math.max(7, Math.round(width * 0.042));
-
-  return (
-    <View
-      style={{
-        width,
-        height,
-        borderRadius,
-        overflow: "hidden",
-        backgroundColor: colors.muted,
-      }}
-    >
-      {coverUri ? (
-        <>
-          {/* Cover image */}
-          <Image
-            source={{ uri: coverUri }}
-            style={StyleSheet.absoluteFillObject}
-            resizeMode="cover"
-          />
-          {/* Realistic book-spine gradient overlay */}
-          <View
-            style={[StyleSheet.absoluteFillObject, { flexDirection: "row" }]}
-            pointerEvents="none"
-          >
-            <View style={{ width: "6%", height: "100%", backgroundColor: "rgba(0,0,0,0.10)" }} />
-            <View
-              style={{ width: "8%", height: "100%", backgroundColor: "rgba(20,20,20,0.20)" }}
-            />
-            <View
-              style={{
-                width: "5%",
-                height: "100%",
-                backgroundColor: "rgba(240,240,240,0.40)",
-              }}
-            />
-            <View
-              style={{
-                width: "18%",
-                height: "100%",
-                backgroundColor: "rgba(215,215,215,0.35)",
-              }}
-            />
-            <View
-              style={{ flex: 1, height: "100%", backgroundColor: "rgba(100,100,100,0.10)" }}
-            />
-          </View>
-          {/* Top highlight */}
-          <View
-            style={{
-              position: "absolute",
-              top: 0,
-              left: 0,
-              right: 0,
-              height: "3%",
-              backgroundColor: "rgba(240,240,240,0.15)",
-            }}
-            pointerEvents="none"
-          />
-          {/* Bottom shadow */}
-          <View
-            style={{
-              position: "absolute",
-              bottom: showStrip ? 2 : 0,
-              left: 0,
-              right: 0,
-              height: "8%",
-              backgroundColor: "rgba(15,15,15,0.15)",
-            }}
-            pointerEvents="none"
-          />
-        </>
-      ) : (
-        /* Fallback: serif gradient — mirrors BookCard style */
-        <View style={{ flex: 1, overflow: "hidden" }}>
-          <View
-            style={{
-              position: "absolute",
-              inset: 0,
-              top: 0,
-              left: 0,
-              right: 0,
-              bottom: 0,
-              backgroundColor: colors.stone100,
-            }}
-          />
-          <View
-            style={{
-              position: "absolute",
-              bottom: 0,
-              left: 0,
-              right: 0,
-              height: "50%",
-              backgroundColor: colors.stone200,
-            }}
-          />
-          <View
-            style={{
-              flex: 1,
-              padding: Math.max(6, width * 0.07),
-              alignItems: "center",
-              justifyContent: "center",
-              zIndex: 1,
-            }}
-          >
-            <View style={{ flex: 1, alignItems: "center", justifyContent: "center" }}>
-              <Text
-                style={{
-                  textAlign: "center",
-                  fontSize,
-                  fontWeight: fontWeight.medium,
-                  fontFamily: "serif",
-                  color: colors.stone500,
-                  lineHeight: Math.round(fontSize * 1.45),
-                }}
-                numberOfLines={4}
-              >
-                {bookTitle || t("reader.untitled")}
-              </Text>
-            </View>
-            <View
-              style={{
-                width: Math.round(width * 0.28),
-                height: 1,
-                backgroundColor: `${colors.stone300}99`,
-                marginVertical: 5,
-              }}
-            />
-            <View style={{ height: "22%", alignItems: "center", justifyContent: "center" }}>
-              <Text
-                style={{
-                  textAlign: "center",
-                  fontSize: subFontSize,
-                  fontFamily: "serif",
-                  color: colors.stone400,
-                }}
-                numberOfLines={1}
-              >
-                {chapterTitle}
-              </Text>
-            </View>
-          </View>
-        </View>
-      )}
-
-      {/* Reading progress strip at bottom edge */}
-      {showStrip && (
-        <View
-          style={{
-            position: "absolute",
-            bottom: 0,
-            left: 0,
-            right: 0,
-            height: 2,
-            backgroundColor: "rgba(0,0,0,0.1)",
-          }}
-        >
-          <View
-            style={{
-              height: "100%" as unknown as number,
-              width: `${pct}%` as unknown as number,
-              backgroundColor: colors.primary,
-              opacity: 0.9,
-            }}
-          />
-        </View>
-      )}
-    </View>
-  );
-}
 
 // ── Props ─────────────────────────────────────────────────────────────────────
 
@@ -1072,757 +867,15 @@ export function TTSPage({
         </View>
       </View>
 
-      {/* ── Engine + Voice Picker Modal ────────────────────────────────────── */}
-      <Modal
-        visible={voicePickerVisible}
-        animationType="slide"
-        transparent
-        onRequestClose={() => setVoicePickerVisible(false)}
-      >
-        <View style={s.voicePickerContainer}>
-        <Pressable
-          style={StyleSheet.absoluteFillObject}
-          onPress={() => setVoicePickerVisible(false)}
+      {onUpdateConfig && (
+        <VoicePickerModal
+          visible={voicePickerVisible}
+          config={config}
+          onClose={() => setVoicePickerVisible(false)}
+          onUpdateConfig={onUpdateConfig}
         />
-        <View style={s.voicePickerSheet}>
-          {/* Handle bar */}
-          <View style={s.voicePickerHandle} />
-
-          {/* Header */}
-          <View style={s.voicePickerHeader}>
-            <Text style={s.voicePickerTitle}>{t("tts.ttsEngine")}</Text>
-          </View>
-
-          {/* Engine selector */}
-          <View style={s.engineSection}>
-            {(["edge", "dashscope", "browser"] as const).map((eng) => {
-              const isActive = config.engine === eng;
-              const label =
-                eng === "edge" ? "Edge TTS" : eng === "dashscope" ? "DashScope" : t("tts.browser");
-              const desc =
-                eng === "edge"
-                  ? "Microsoft · 多语言"
-                  : eng === "dashscope"
-                    ? "阿里云通义 · 中文优化"
-                    : "系统内置 · 免费";
-              return (
-                <TouchableOpacity
-                  key={eng}
-                  style={[s.engineRow, isActive && s.engineRowActive]}
-                  onPress={() => onUpdateConfig?.({ engine: eng })}
-                  activeOpacity={0.7}
-                >
-                  <View style={s.engineRowLeft}>
-                    <Text style={[s.engineRowLabel, isActive && s.engineRowLabelActive]}>
-                      {label}
-                    </Text>
-                    <Text style={s.engineRowDesc}>{desc}</Text>
-                  </View>
-                  {isActive && (
-                    <View style={s.engineCheckmark}>
-                      <Text style={s.engineCheckmarkTxt}>✓</Text>
-                    </View>
-                  )}
-                </TouchableOpacity>
-              );
-            })}
-          </View>
-
-          {/* Divider + voice section title */}
-          {config.engine !== "browser" && (
-            <View style={s.voicePickerHeader}>
-              <Text style={s.voicePickerTitle}>{t("tts.selectVoice")}</Text>
-            </View>
-          )}
-
-          <ScrollView
-            style={s.voicePickerList}
-            showsVerticalScrollIndicator={false}
-          >
-            {/* DashScope voices */}
-            {config.engine === "dashscope" &&
-              DASHSCOPE_VOICES.map((v) => {
-                const isSelected = config.dashscopeVoice === v.id;
-                return (
-                  <TouchableOpacity
-                    key={v.id}
-                    style={[s.voiceItem, isSelected && s.voiceItemSelected]}
-                    onPress={() => {
-                      onUpdateConfig?.({ dashscopeVoice: v.id });
-                      setVoicePickerVisible(false);
-                    }}
-                    activeOpacity={0.7}
-                  >
-                    <Text style={[s.voiceItemTxt, isSelected && s.voiceItemTxtSelected]}>
-                      {v.label}
-                    </Text>
-                    {isSelected && (
-                      <Text style={s.voiceItemCheck}>✓</Text>
-                    )}
-                  </TouchableOpacity>
-                );
-              })}
-
-            {/* Edge TTS voices — grouped by language, zh-* first */}
-            {config.engine === "edge" &&
-              (() => {
-                const grouped = EDGE_TTS_VOICES.reduce<Record<string, typeof EDGE_TTS_VOICES>>(
-                  (acc, v) => {
-                    (acc[v.lang] ??= []).push(v);
-                    return acc;
-                  },
-                  {},
-                );
-                const langs = Object.keys(grouped).sort((a, b) => {
-                  const aZh = a.startsWith("zh") ? -1 : 0;
-                  const bZh = b.startsWith("zh") ? -1 : 0;
-                  return aZh - bZh || a.localeCompare(b);
-                });
-                return langs.map((lang) => (
-                  <View key={lang}>
-                    <View style={s.voiceLangHeader}>
-                      <Text style={s.voiceLangTxt}>{lang}</Text>
-                    </View>
-                    {grouped[lang].map((v) => {
-                      const isSelected = config.edgeVoice === v.id;
-                      return (
-                        <TouchableOpacity
-                          key={v.id}
-                          style={[s.voiceItem, isSelected && s.voiceItemSelected]}
-                          onPress={() => {
-                            onUpdateConfig?.({ edgeVoice: v.id });
-                            setVoicePickerVisible(false);
-                          }}
-                          activeOpacity={0.7}
-                        >
-                          <Text style={[s.voiceItemTxt, isSelected && s.voiceItemTxtSelected]}>
-                            {v.name}
-                          </Text>
-                          {isSelected && (
-                            <Text style={s.voiceItemCheck}>✓</Text>
-                          )}
-                        </TouchableOpacity>
-                      );
-                    })}
-                  </View>
-                ));
-              })()}
-
-            {/* Browser — no selectable voices */}
-            {config.engine === "browser" && (
-              <View style={s.voiceBrowserNote}>
-                <Text style={s.voiceBrowserNoteTxt}>
-                  {t("tts.browserVoiceNote")}
-                </Text>
-              </View>
-            )}
-          </ScrollView>
-
-          {/* Cancel button */}
-          <TouchableOpacity
-            style={s.voicePickerCancel}
-            onPress={() => setVoicePickerVisible(false)}
-            activeOpacity={0.7}
-          >
-            <Text style={s.voicePickerCancelTxt}>{t("common.cancel")}</Text>
-          </TouchableOpacity>
-        </View>
-        </View>
-      </Modal>
+      )}
     </Modal>
   );
 }
 
-// ── Styles ────────────────────────────────────────────────────────────────────
-
-const makeStyles = (colors: ThemeColors) =>
-  StyleSheet.create({
-    screen: {
-      flex: 1,
-      backgroundColor: colors.background,
-    },
-    content: {
-      flex: 1,
-      backgroundColor: colors.background,
-    },
-
-    // ── Pager (native horizontal ScrollView) ─────────────────────────────
-    pager: {
-      flex: 1,
-    },
-    pagerContent: {
-      // contentContainerStyle — pages define their own width via s.page
-    },
-    page: {
-      width: SW,
-      flex: 1,
-      flexDirection: "column",
-      backgroundColor: colors.background,
-    },
-
-    // ── Top bar — sits above the pager, never scrolls ─────────────────────
-    topBar: {
-      flexDirection: "row",
-      alignItems: "center",
-      justifyContent: "space-between",
-      paddingHorizontal: 20,
-      paddingTop: 52,
-      paddingBottom: 4,
-    },
-    iconBtn: {
-      width: 40,
-      height: 36,
-      alignItems: "center",
-      justifyContent: "center",
-    },
-    statusPill: {
-      flexDirection: "row",
-      alignItems: "center",
-      gap: 5,
-      paddingHorizontal: 12,
-      paddingVertical: 5,
-      borderRadius: 999,
-      backgroundColor: withOpacity(colors.primary, 0.1),
-    },
-    statusTxt: {
-      fontSize: 11,
-      fontWeight: fontWeight.semibold,
-      color: colors.primary,
-      textTransform: "uppercase",
-      letterSpacing: 0.5,
-    },
-    // ── Cover section ─────────────────────────────────────────────────────
-    coverSection: {
-      flex: 1,
-      alignItems: "center",
-      justifyContent: "center",
-      paddingVertical: 8,
-    },
-    /**
-     * Ambient glow effect — intentionally NO backgroundColor.
-     * Without a fill, this is invisible as a shape; the shadow
-     * properties alone produce the soft halo on iOS (shadowColor /
-     * shadowRadius). On Android the elevation tint from the cover
-     * itself provides the same effect.
-     */
-    glow: {
-      position: "absolute",
-      width: COVER_W + 60,
-      height: COVER_H + 40,
-      borderRadius: (COVER_W + 60) / 2,
-      // NO backgroundColor — this was the source of the "圆环" ring artifact
-      shadowColor: colors.primary,
-      shadowOffset: { width: 0, height: 0 },
-      shadowOpacity: 0.22,
-      shadowRadius: 64,
-    },
-    coverShadow: {
-      shadowColor: "#000",
-      shadowOffset: { width: 0, height: 18 },
-      shadowOpacity: 0.28,
-      shadowRadius: 32,
-      elevation: 16,
-    },
-
-    // ── Book info below cover ─────────────────────────────────────────────
-    bookInfo: {
-      alignItems: "center",
-      paddingHorizontal: 36,
-      paddingTop: 4,
-      paddingBottom: 8,
-    },
-    bookTitle: {
-      fontSize: 18,
-      fontWeight: fontWeight.bold,
-      color: colors.foreground,
-      textAlign: "center",
-      lineHeight: 26,
-    },
-    bookChapter: {
-      fontSize: 14,
-      color: colors.mutedForeground,
-      textAlign: "center",
-      lineHeight: 20,
-      marginTop: 2,
-    },
-
-    // ── Progress bar ──────────────────────────────────────────────────────
-    progress: {
-      paddingHorizontal: 24,
-      paddingBottom: 8,
-    },
-    progressTrack: {
-      height: 3,
-      borderRadius: 999,
-      backgroundColor: withOpacity(colors.border, 0.45),
-      overflow: "hidden",
-    },
-    progressFill: {
-      height: "100%" as unknown as number,
-      borderRadius: 999,
-      backgroundColor: colors.primary,
-    },
-    progressRow: {
-      flexDirection: "row",
-      justifyContent: "space-between",
-      marginTop: 5,
-    },
-    progressTxt: {
-      fontSize: 10,
-      color: colors.mutedForeground,
-    },
-
-    // ── Transport controls ─────────────────────────────────────────────────
-    controls: {
-      flexDirection: "row",
-      alignItems: "center",
-      justifyContent: "center",
-      gap: 20,
-      paddingVertical: 4,
-      paddingBottom: 10,
-    },
-    ctrlBtn: {
-      width: 46,
-      height: 46,
-      borderRadius: 23,
-      alignItems: "center",
-      justifyContent: "center",
-      backgroundColor: withOpacity(colors.muted, 0.7),
-    },
-    /** Smaller secondary button for prev/next chapter */
-    ctrlBtnSm: {
-      width: 38,
-      height: 38,
-      borderRadius: 19,
-      alignItems: "center",
-      justifyContent: "center",
-      backgroundColor: withOpacity(colors.muted, 0.5),
-    },
-    ctrlBtnDisabled: {
-      opacity: 0.3,
-    },
-    playBtn: {
-      width: 68,
-      height: 68,
-      borderRadius: 34,
-      alignItems: "center",
-      justifyContent: "center",
-      backgroundColor: colors.primary,
-      shadowColor: colors.primary,
-      shadowOffset: { width: 0, height: 6 },
-      shadowOpacity: 0.35,
-      shadowRadius: 16,
-      elevation: 10,
-    },
-
-    // ── Settings row ──────────────────────────────────────────────────────
-    settings: {
-      flexDirection: "row",
-      alignItems: "center",
-      marginHorizontal: 20,
-      marginBottom: 8,
-      borderRadius: 16,
-      backgroundColor: withOpacity(colors.muted, 0.5),
-      paddingHorizontal: 18,
-      paddingVertical: 12,
-    },
-    settingGroup: {
-      flex: 1,
-      flexDirection: "row",
-      alignItems: "center",
-      justifyContent: "space-between",
-    },
-    settingLbl: {
-      fontSize: 11,
-      color: colors.mutedForeground,
-      fontWeight: fontWeight.medium,
-    },
-    stepper: {
-      flexDirection: "row",
-      alignItems: "center",
-      gap: 7,
-    },
-    stepBtn: {
-      width: 24,
-      height: 24,
-      borderRadius: 12,
-      alignItems: "center",
-      justifyContent: "center",
-      backgroundColor: colors.card,
-      borderWidth: 1,
-      borderColor: withOpacity(colors.border, 0.6),
-    },
-    stepVal: {
-      fontSize: 12,
-      color: colors.foreground,
-      fontWeight: fontWeight.semibold,
-      minWidth: 30,
-      textAlign: "center",
-    },
-    settingDiv: {
-      width: 1,
-      height: 20,
-      backgroundColor: withOpacity(colors.border, 0.5),
-      marginHorizontal: 10,
-    },
-    contBtn: {
-      borderRadius: 999,
-      paddingHorizontal: 9,
-      paddingVertical: 4,
-      backgroundColor: withOpacity(colors.background, 0.7),
-      borderWidth: 1,
-      borderColor: withOpacity(colors.border, 0.5),
-    },
-    contBtnOn: {
-      backgroundColor: withOpacity(colors.primary, 0.12),
-      borderColor: withOpacity(colors.primary, 0.3),
-    },
-    contBtnTxt: {
-      fontSize: 10,
-      fontWeight: fontWeight.semibold,
-      color: colors.mutedForeground,
-    },
-    contBtnTxtOn: {
-      color: colors.primary,
-    },
-
-    // ── Bottom row ────────────────────────────────────────────────────────
-    bottom: {
-      alignItems: "center",
-      justifyContent: "center",
-      paddingBottom: 34,
-      paddingTop: 4,
-    },
-    /** Bottom strip — continuous toggle on left, chips on right */
-    bottomStrip: {
-      flexDirection: "row",
-      alignItems: "center",
-      justifyContent: "space-between",
-      paddingHorizontal: 20,
-      width: "100%" as unknown as number,
-      gap: 8,
-    },
-    bottomStripLeft: {
-      flexDirection: "row",
-      alignItems: "center",
-      gap: 8,
-      flexShrink: 1,
-    },
-    returnBtn: {
-      borderRadius: 999,
-      paddingHorizontal: 10,
-      paddingVertical: 4,
-      backgroundColor: withOpacity(colors.primary, 0.1),
-      borderWidth: 1,
-      borderColor: withOpacity(colors.primary, 0.22),
-    },
-    returnBtnTxt: {
-      fontSize: 10,
-      fontWeight: fontWeight.semibold,
-      color: colors.primary,
-    },
-    /** Animated page indicator — inner container */
-    pageIndicator: {
-      // PILL_W(18) + GAP(5) + DOT(6) = 29 — all children are absolutely positioned
-      width: 29,
-      height: 6,
-    },
-    dot: {
-      width: 6,
-      height: 6,
-      borderRadius: 3,
-      backgroundColor: withOpacity(colors.mutedForeground, 0.65),
-    },
-    /** Helper — makes a dot absolutely positioned within pageIndicator */
-    dotAbs: {
-      position: "absolute",
-      top: 0,
-    },
-    /** Animated pill — absolutely positioned at left:0, translateX 0→11 */
-    dotPill: {
-      position: "absolute",
-      top: 0,
-      left: 0,
-      height: 6,
-      borderRadius: 3,
-      backgroundColor: colors.primary,
-    },
-    chips: {
-      flexDirection: "row",
-      alignItems: "center",
-      gap: 5,
-    },
-    chip: {
-      paddingHorizontal: 9,
-      paddingVertical: 3,
-      borderRadius: 999,
-      backgroundColor: withOpacity(colors.muted, 0.9),
-      maxWidth: 130,
-    },
-    chipTxt: {
-      fontSize: 10,
-      color: colors.mutedForeground,
-      fontWeight: fontWeight.medium,
-    },
-
-    // ── Lyrics view ───────────────────────────────────────────────────────
-
-    /** Thumbnail + book meta side by side below top bar */
-    lyricHeaderRow: {
-      flexDirection: "row",
-      alignItems: "center",
-      paddingHorizontal: 22,
-      paddingVertical: 10,
-      gap: 14,
-    },
-    thumbShadowWrap: {
-      shadowColor: "#000",
-      shadowOffset: { width: 0, height: 4 },
-      shadowOpacity: 0.18,
-      shadowRadius: 8,
-      elevation: 6,
-    },
-    lyricHeaderMeta: {
-      flex: 1,
-    },
-    lyricBookName: {
-      fontSize: 15,
-      fontWeight: fontWeight.semibold,
-      color: colors.foreground,
-      lineHeight: 22,
-    },
-    lyricChapterName: {
-      fontSize: 12,
-      color: colors.mutedForeground,
-      lineHeight: 18,
-      marginTop: 3,
-    },
-
-    /** Karaoke text area — centered, fills remaining space */
-    lyricArea: {
-      flex: 1,
-      paddingHorizontal: 24,
-      paddingTop: 8,
-      paddingBottom: 16,
-      alignSelf: "stretch",
-    },
-    lyricList: {
-      flex: 1,
-      alignSelf: "stretch",
-    },
-    lyricListContent: {
-      justifyContent: "center",
-    },
-    lyricInner: {
-      gap: 8,
-    },
-    lyricLinePressable: {
-      paddingHorizontal: 14,
-      paddingVertical: 10,
-      borderRadius: 18,
-    },
-    lyricLinePressableActive: {
-      backgroundColor: withOpacity(colors.foreground, 0.05),
-    },
-    lyricLine: {
-      fontSize: 16,
-      fontWeight: fontWeight.medium,
-      color: colors.foreground,
-      textAlign: "center",
-      lineHeight: 27,
-      opacity: 0.32,
-    },
-    lyricLineActive: {
-      fontSize: 22,
-      fontWeight: fontWeight.bold,
-      lineHeight: 34,
-      opacity: 1,
-      color: colors.foreground,
-    },
-    lyricLinePast: {
-      opacity: 0.58,
-    },
-    lyricContextLabel: {
-      fontSize: 11,
-      color: colors.mutedForeground,
-      textAlign: "center",
-      marginBottom: 8,
-    },
-    lyricSectionDivider: {
-      height: 1,
-      backgroundColor: colors.mutedForeground,
-      opacity: 0.2,
-      marginVertical: 10,
-      marginHorizontal: 16,
-    },
-    lyricHintLabel: {
-      fontSize: 11,
-      color: colors.mutedForeground,
-      textAlign: "center",
-      marginTop: 8,
-    },
-    lyricLoadingFooter: {
-      alignItems: "center",
-      paddingVertical: 16,
-      opacity: 0.5,
-    },
-    lyricActive: {
-      fontSize: 18,
-      fontWeight: fontWeight.bold,
-      color: colors.foreground,
-      textAlign: "center",
-      lineHeight: 28,
-    },
-
-    // ── Engine + Voice picker bottom sheet ───────────────────────────────
-    voicePickerContainer: {
-      flex: 1,
-      justifyContent: "flex-end",
-      backgroundColor: "rgba(0,0,0,0.45)",
-    },
-    voicePickerSheet: {
-      backgroundColor: colors.background,
-      borderTopLeftRadius: 24,
-      borderTopRightRadius: 24,
-      paddingBottom: 34,
-      maxHeight: SH * 0.82,
-    },
-    voicePickerHandle: {
-      width: 36,
-      height: 4,
-      borderRadius: 2,
-      backgroundColor: colors.border,
-      alignSelf: "center",
-      marginTop: 12,
-      marginBottom: 4,
-    },
-    voicePickerHeader: {
-      paddingHorizontal: 20,
-      paddingTop: 14,
-      paddingBottom: 8,
-    },
-    voicePickerTitle: {
-      fontSize: 11,
-      fontWeight: fontWeight.semibold,
-      color: colors.mutedForeground,
-      textTransform: "uppercase",
-      letterSpacing: 0.9,
-    },
-    // Engine selector rows
-    engineSection: {
-      marginHorizontal: 16,
-      marginBottom: 4,
-      borderRadius: 14,
-      overflow: "hidden",
-      borderWidth: StyleSheet.hairlineWidth,
-      borderColor: colors.border,
-    },
-    engineRow: {
-      flexDirection: "row",
-      alignItems: "center",
-      justifyContent: "space-between",
-      paddingHorizontal: 16,
-      paddingVertical: 13,
-      backgroundColor: colors.background,
-      borderBottomWidth: StyleSheet.hairlineWidth,
-      borderBottomColor: withOpacity(colors.border, 0.5),
-    },
-    engineRowActive: {
-      backgroundColor: withOpacity(colors.primary, 0.06),
-    },
-    engineRowLeft: {
-      flex: 1,
-    },
-    engineRowLabel: {
-      fontSize: 15,
-      fontWeight: fontWeight.medium,
-      color: colors.foreground,
-    },
-    engineRowLabelActive: {
-      color: colors.primary,
-      fontWeight: fontWeight.semibold,
-    },
-    engineRowDesc: {
-      fontSize: 11,
-      color: colors.mutedForeground,
-      marginTop: 2,
-    },
-    engineCheckmark: {
-      width: 24,
-      height: 24,
-      borderRadius: 12,
-      backgroundColor: colors.primary,
-      alignItems: "center",
-      justifyContent: "center",
-      marginLeft: 12,
-    },
-    engineCheckmarkTxt: {
-      fontSize: 13,
-      color: "#fff",
-      fontWeight: fontWeight.bold,
-    },
-    // Voice list
-    voicePickerList: {
-      flexGrow: 0,
-    },
-    voiceLangHeader: {
-      paddingHorizontal: 20,
-      paddingVertical: 6,
-      backgroundColor: withOpacity(colors.muted, 0.6),
-    },
-    voiceLangTxt: {
-      fontSize: 10,
-      fontWeight: fontWeight.semibold,
-      color: colors.mutedForeground,
-      textTransform: "uppercase",
-      letterSpacing: 0.7,
-    },
-    voiceItem: {
-      flexDirection: "row",
-      alignItems: "center",
-      justifyContent: "space-between",
-      paddingHorizontal: 20,
-      paddingVertical: 15,
-      borderBottomWidth: StyleSheet.hairlineWidth,
-      borderBottomColor: withOpacity(colors.border, 0.4),
-    },
-    voiceItemSelected: {
-      backgroundColor: withOpacity(colors.primary, 0.06),
-    },
-    voiceItemTxt: {
-      fontSize: 15,
-      color: colors.foreground,
-    },
-    voiceItemTxtSelected: {
-      fontWeight: fontWeight.semibold,
-      color: colors.primary,
-    },
-    voiceItemCheck: {
-      fontSize: 15,
-      fontWeight: fontWeight.bold,
-      color: colors.primary,
-    },
-    voiceBrowserNote: {
-      padding: 28,
-      alignItems: "center",
-    },
-    voiceBrowserNoteTxt: {
-      fontSize: 14,
-      color: colors.mutedForeground,
-      textAlign: "center",
-      lineHeight: 20,
-    },
-    voicePickerCancel: {
-      marginHorizontal: 16,
-      marginTop: 12,
-      paddingVertical: 15,
-      borderRadius: 14,
-      backgroundColor: colors.muted,
-      alignItems: "center",
-    },
-    voicePickerCancelTxt: {
-      fontSize: 15,
-      fontWeight: fontWeight.semibold,
-      color: colors.foreground,
-    },
-  });
