+++ packages/app-expo/src/screens/stats/stats-styles.ts
@@ -1,0 +1,231 @@
+import { StyleSheet } from "react-native";
+import {
+  type ThemeColors,
+  fontSize,
+  fontWeight,
+  radius,
+  withOpacity,
+} from "@/styles/theme";
+
+export const makeStyles = (colors: ThemeColors) =>
+  StyleSheet.create({
+    container: { flex: 1, backgroundColor: colors.background },
+    loadingWrap: { flex: 1, alignItems: "center", justifyContent: "center" },
+    header: {
+      flexDirection: "row",
+      alignItems: "center",
+      justifyContent: "space-between",
+      paddingHorizontal: 12,
+      paddingVertical: 10,
+      borderBottomWidth: 0.5,
+      borderBottomColor: colors.border,
+    },
+    backBtn: {
+      width: 36,
+      height: 36,
+      borderRadius: radius.full,
+      alignItems: "center",
+      justifyContent: "center",
+    },
+    headerTitle: {
+      fontSize: fontSize.lg,
+      fontWeight: fontWeight.semibold,
+      color: colors.foreground,
+    },
+    scrollContent: { padding: 16 },
+
+    statsGrid: { flexDirection: "row", flexWrap: "wrap", gap: 12, marginBottom: 12 },
+    statCard: {
+      width: "47%",
+      backgroundColor: colors.card,
+      borderRadius: radius.xl,
+      borderWidth: 0.5,
+      borderColor: colors.border,
+      padding: 14,
+    },
+    statCardHeader: {
+      flexDirection: "row",
+      alignItems: "center",
+      justifyContent: "space-between",
+      marginBottom: 6,
+    },
+    statCardTitle: { fontSize: fontSize.xs, color: colors.mutedForeground },
+    statCardBody: { flexDirection: "row", alignItems: "baseline", gap: 4 },
+    statCardValue: {
+      fontSize: fontSize["2xl"],
+      fontWeight: fontWeight.bold,
+      color: colors.foreground,
+    },
+    statCardUnit: { fontSize: fontSize.sm, color: colors.mutedForeground },
+
+    section: { marginBottom: 12 },
+    sectionCard: {
+      backgroundColor: colors.card,
+      borderRadius: radius.xl,
+      borderWidth: 0.5,
+      borderColor: colors.border,
+      padding: 14,
+    },
+    sectionCardTitle: {
+      fontSize: fontSize.sm,
+      fontWeight: fontWeight.medium,
+      color: colors.mutedForeground,
+      marginBottom: 12,
+    },
+
+    chartHeaderRow: {
+      flexDirection: "row",
+      alignItems: "center",
+      justifyContent: "space-between",
+      marginBottom: 12,
+    },
+    chartHeaderLabel: {
+      fontSize: fontSize.sm,
+      fontWeight: fontWeight.medium,
+      color: colors.mutedForeground,
+    },
+    toggleRow: {
+      flexDirection: "row",
+      borderRadius: radius.md,
+      borderWidth: 0.5,
+      borderColor: colors.border,
+      backgroundColor: colors.muted,
+      padding: 2,
+    },
+    toggleBtn: { paddingHorizontal: 8, paddingVertical: 3, borderRadius: radius.sm },
+    toggleBtnActive: { backgroundColor: colors.background },
+    toggleBtnText: { fontSize: 12, fontWeight: fontWeight.medium, color: colors.mutedForeground },
+    toggleBtnTextActive: { color: colors.foreground },
+
+    barControlsRow: {
+      flexDirection: "row",
+      alignItems: "center",
+      justifyContent: "space-between",
+      marginBottom: 12,
+    },
+    periodNav: { flexDirection: "row", alignItems: "center", gap: 2 },
+    periodNavBtn: { padding: 4, borderRadius: radius.sm },
+    periodLabel: {
+      fontSize: 12,
+      fontWeight: fontWeight.medium,
+      color: colors.mutedForeground,
+      minWidth: 80,
+      textAlign: "center",
+    },
+
+    heatmapLegend: {
+      flexDirection: "row",
+      alignItems: "center",
+      justifyContent: "flex-end",
+      gap: 4,
+      marginTop: 10,
+    },
+    legendText: { fontSize: 9, color: colors.mutedForeground },
+    legendCell: { width: 10, height: 10, borderRadius: 2 },
+
+    barChartWrap: { height: 180 },
+    barChartContent: { alignItems: "flex-end", gap: 4, paddingBottom: 4 },
+    barCol: { alignItems: "center", width: 28 },
+    barTrack: { justifyContent: "flex-end", width: 16 },
+    barFill: { width: 16, borderRadius: 4 },
+    barLabel: { fontSize: 8, color: colors.mutedForeground, marginTop: 4 },
+    barChartEmpty: { height: 120, alignItems: "center", justifyContent: "center" },
+    barChartEmptyText: { fontSize: fontSize.xs, color: colors.mutedForeground },
+    tooltip: {
+      position: "absolute",
+      backgroundColor: colors.card,
+      paddingHorizontal: 6,
+      paddingVertical: 2,
+      borderRadius: 4,
+      borderWidth: 0.5,
+      borderColor: colors.border,
+    },
+    tooltipText: { fontSize: 9, color: colors.cardForeground, fontWeight: "500" },
+
+    trendContent: { alignItems: "flex-end", gap: 1 },
+    trendCol: { alignItems: "center", width: 10 },
+    trendTrack: { justifyContent: "flex-end", width: 6 },
+    trendBar: { width: 6, borderTopLeftRadius: 2, borderTopRightRadius: 2 },
+    trendLabel: { fontSize: 7, color: colors.mutedForeground, marginTop: 2 },
+
+    periodBooksEmpty: {
+      fontSize: fontSize.xs,
+      color: colors.mutedForeground,
+      textAlign: "center",
+      paddingVertical: 16,
+    },
+    bookRow: {
+      flexDirection: "row",
+      alignItems: "center",
+      gap: 10,
+      paddingVertical: 6,
+      paddingHorizontal: 6,
+      borderRadius: radius.lg,
+    },
+    bookCoverPlaceholder: {
+      width: 28,
+      height: 40,
+      borderRadius: radius.sm,
+      backgroundColor: colors.muted,
+      alignItems: "center",
+      justifyContent: "center",
+    },
+    bookCover: {
+      width: 28,
+      height: 40,
+      borderRadius: radius.sm,
+      backgroundColor: colors.muted,
+    },
+    bookCoverLetter: { fontSize: 12, color: colors.mutedForeground },
+    bookInfo: { flex: 1, gap: 4 },
+    bookTitleRow: {
+      flexDirection: "row",
+      alignItems: "center",
+      justifyContent: "space-between",
+      gap: 4,
+    },
+    bookTitle: {
+      fontSize: fontSize.xs,
+      fontWeight: fontWeight.medium,
+      color: colors.foreground,
+      flex: 1,
+    },
+    bookAuthor: { fontSize: 11, color: colors.mutedForeground },
+    bookTime: { fontSize: 12, color: colors.mutedForeground },
+    progressRow: { flexDirection: "row", alignItems: "center", gap: 6 },
+    progressTrack: {
+      flex: 1,
+      height: 4,
+      borderRadius: 2,
+      backgroundColor: colors.muted,
+      overflow: "hidden",
+    },
+    progressFill: { height: "100%", borderRadius: 2, backgroundColor: colors.emerald },
+    progressPercent: { fontSize: 11, color: colors.mutedForeground },
+
+    streakCard: {
+      flexDirection: "row",
+      alignItems: "center",
+      gap: 10,
+      backgroundColor: colors.card,
+      borderRadius: radius.xl,
+      borderWidth: 0.5,
+      borderColor: colors.border,
+      padding: 12,
+    },
+    streakIconWrap: {
+      width: 32,
+      height: 32,
+      borderRadius: radius.lg,
+      backgroundColor: withOpacity(colors.amber, 0.1),
+      alignItems: "center",
+      justifyContent: "center",
+    },
+    streakInfo: { gap: 2 },
+    streakLabel: {
+      fontSize: fontSize.xs,
+      fontWeight: fontWeight.semibold,
+      color: colors.foreground,
+    },
+    streakDesc: { fontSize: 12, color: colors.mutedForeground },
+  });
