+++ packages/app-expo/src/screens/stats/stats-utils.ts
@@ -1,0 +1,29 @@
+/** Utility functions shared across stats components. */
+
+export function formatTime(minutes: number): string {
+  if (minutes < 60) return `${Math.round(minutes)}m`;
+  const h = Math.floor(minutes / 60);
+  const m = Math.round(minutes % 60);
+  return m > 0 ? `${h}h${m}m` : `${h}h`;
+}
+
+export function formatDate(dateStr: string): string {
+  const [, m, d] = dateStr.split("-");
+  return `${Number.parseInt(m)}/${Number.parseInt(d)}`;
+}
+
+export function getWeekStart(date: Date): Date {
+  const d = new Date(date);
+  const day = d.getDay();
+  const diff = day === 0 ? -6 : 1 - day;
+  d.setDate(d.getDate() + diff);
+  d.setHours(0, 0, 0, 0);
+  return d;
+}
+
+export function getWeekEnd(weekStart: Date): Date {
+  const d = new Date(weekStart);
+  d.setDate(d.getDate() + 6);
+  d.setHours(23, 59, 59, 999);
+  return d;
+}
